requirements.txt
cvxopt==1.3.2
matplotlib==3.7.2
numpy==1.24.4
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
opencv_python_headless==4.8.0.74
pandas==1.4.4
Pillow==10.2.0
scikit_image==0.19.3
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
setuptools==65.5.0
torch==2.2.1+cu121
torch==1.13.1+cu117
torchvision==0.17.1+cu121